fix build on x86_64
