+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "linux-syscall",
     "linux-loader",
     "kernel-hal-unix",
-    "kernel-hal-bare",
     "kernel-hal",
 ]
